T0.0.1A - Builds Successfully
This branch is a temporary branch while we port the 2019 Robot code to the 2020 framework changes. This branch serves two purposes.
A) It allows us to learn the changes between the 2019 and 2020 ways of doing things
B) It will serve as a foundation for what should be transferred to the 2020 master branch

TODO:
Fix deprecated gradle features warning
Fix deprecated Command usage
Fix deprecated PIDCommand usage