S.6.4.0S - Proportional Intake Percentage
I have added a method to the intake subsystem that is supposed to calculate the speed that the intake should be set to based on the speed of the robot's drivetrain. Getter methods have also been added to the Drive subsystem in order to get the left, right, and average velocities.